Add admin to payment system when createAdmin is complete

+++ common/src/main/java/com/seasidechachacha/common/payment/NewUserRequest.java
@@ -9,7 +9,7 @@ public class NewUserRequest implements Serializable {
      */
     double deposit;
 
-    public NewUserRequest(String userId, int deposit) {
+    public NewUserRequest(String userId, double deposit) {
         this.userId = userId;
         this.deposit = deposit;
     }
